# Improve unique CPU performance for returning counts