--- a/aten/src/ATen/native/Unique.cpp
+++ b/aten/src/ATen/native/Unique.cpp
@@ -2,7 +2,6 @@
 
 #include <ATen/ATen.h>
 #include <ATen/Dispatch.h>
-#include <ATen/Parallel.h>
 
 #include <set>
 #include <tuple>
@@ -47,30 +46,24 @@ std::tuple<Tensor, Tensor, Tensor> unique_cpu_template(
     for (int i = 0; i < output.numel(); ++i) {
       inverse_map[output_data[i]] = i;
     }
-    at::parallel_for(0, numel, 1, [&](int64_t i_begin, int64_t i_end) {
-      for(int64_t i = i_begin; i < i_end; i++) {
+    for(int64_t i = 0; i < numel; i++) {
         inverse_indices_data[i] = inverse_map[input_data[i]];
-      }
-    });
+    }
     if (return_counts) {
-      std::unordered_map<scalar_t, std::atomic_int64_t> counts_map;
+      std::unordered_map<scalar_t, int64_t> counts_map;
       counts_map.reserve(output.numel());
       for (int i = 0; i < output.numel(); ++i) {
         counts_map[output_data[i]] = 0;
       }
-      at::parallel_for(0, numel, 1, [&](int64_t i_begin, int64_t i_end) {
-        for(int64_t i = i_begin; i < i_end; i++) {
-          counts_map[input_data[i]] += 1;
-        }
-      });
+      for(int64_t i = 0; i < numel; i++) {
+        counts_map[input_data[i]] += 1;
+      }
       counts.resize_(output.sizes());
       counts.fill_(0);
       int64_t *counts_data = counts.data<int64_t>();
-      at::parallel_for(0, output.numel(), 1, [&](int64_t i_begin, int64_t i_end) {
-        for(int64_t i = i_begin; i < i_end; i++) {
-          counts_data[i] = counts_map[output_data[i]];
-        }
-      });
+      for(int64_t i = 0; i < output.numel(); i++) {
+        counts_data[i] = counts_map[output_data[i]];
+      }
     }
   }
   return std::make_tuple(output, inverse_indices, counts);