--- a/aten/src/ATen/native/Unique.cpp
+++ b/aten/src/ATen/native/Unique.cpp
@@ -46,8 +46,8 @@ std::tuple<Tensor, Tensor, Tensor> unique_cpu_template(
     for (int i = 0; i < output.numel(); ++i) {
       inverse_map[output_data[i]] = i;
     }
-    for(int64_t i = 0; i < numel; i++) {
-        inverse_indices_data[i] = inverse_map[input_data[i]];
+    for(int64_t i = 0; i < numel; ++i) {
+       inverse_indices_data[i] = inverse_map[input_data[i]];
     }
     if (return_counts) {
       std::unordered_map<scalar_t, int64_t> counts_map;