--- a/aten/src/ATen/native/Unique.cpp
+++ b/aten/src/ATen/native/Unique.cpp
@@ -43,7 +43,7 @@ std::tuple<Tensor, Tensor, Tensor> unique_cpu_template(
     int64_t* inverse_indices_data = inverse_indices.data<int64_t>();
     std::unordered_map<scalar_t, int64_t> inverse_map;
     inverse_map.reserve(output.numel());
-    for (int i = 0; i < output.numel(); ++i) {
+    for (int64_t i = 0; i < output.numel(); ++i) {
       inverse_map[output_data[i]] = i;
     }
     for(int64_t i = 0; i < numel; ++i) {
@@ -52,7 +52,7 @@ std::tuple<Tensor, Tensor, Tensor> unique_cpu_template(
     if (return_counts) {
       std::unordered_map<scalar_t, int64_t> counts_map;
       counts_map.reserve(output.numel());
-      for (int i = 0; i < output.numel(); ++i) {
+      for (int64_t i = 0; i < output.numel(); ++i) {
         counts_map[output_data[i]] = 0;
       }
       for(int64_t i = 0; i < numel; i++) {
@@ -184,7 +184,7 @@ std::tuple<Tensor, Tensor, Tensor> _unique_dim_cpu_template(
   Tensor input_sorted;
   if (!consecutive) {
     input_sorted = at::empty(input_flat.sizes(), input_flat.options());
-    for (int i = 0; i < indices.size(); ++i) {
+    for (int64_t i = 0; i < indices.size(); ++i) {
       input_sorted[i] = input_flat[indices[i]];
     }
   } else {